adds pages directory and moves app page views to new directory
