+++ client/src/index.js
@@ -3,7 +3,7 @@ import React from 'react';
 import ReactDOM, { createRoot } from 'react-dom/client';
 
 /*========== INTERNAL MODULES ==========*/
-import App from './components/App.jsx';
+import App from './pages/App.jsx';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root')
